flush executor
